+++ test/server/github-tokens.spec.js
@@ -2,42 +2,24 @@ var assert = require('chai').assert;
 var expect = require('chai').expect;
 var sinon = require('sinon');
 var config = require('../../server/config')
-var unirest = require('unirest')
 var jwt = require('jsonwebtoken');
 var fs = require('fs');
 
-var GithubApp = require("../../server/github-app")
+var GithubServiceApi = require("../../server/github-service-api")
+var GithubApp = require("../../server/github-tokens")
 
 describe('Github Apps', function() {
     var uut;
 
     var sandbox = sinon.createSandbox();
     var token = "tokentestcontent"
-    var unirestMock, mockResponse
 
     beforeEach(function() {
-        mockResponse = {
-            ok: true,
-            body: {
-                token: token
-            }
-        }
-        
-        unirestMock = {
-            headers: sinon.stub().returns(unirestMock),
-            end: function(cb) {
-                cb(mockResponse)
-            }
-        }
-
         // mock private key
         sandbox.stub(fs, "readFileSync")
 
         // mock jwt signing
         sandbox.stub(jwt, "sign")
-
-        sandbox.stub(unirest, "post").returns(unirestMock);
-        sandbox.stub(unirest, "get").returns(unirestMock)
         
         uut = GithubApp()
     });
@@ -96,8 +78,13 @@ describe('Github Apps', function() {
     })
 
     describe('Bearer Tokens', function() {
+        var token = "testtoken"
 
         it('should return token on success', function(complete) {
+            sandbox.stub(GithubServiceApi, "requestAccessTokens").callsFake((id, jwt, fn) => {
+                fn(token)
+            })
+
             uut.createBearer(0, function(bearer, err) {
                 expect(bearer.token).to.be.equal(token)
                 expect(err).to.be.null
@@ -106,6 +93,10 @@ describe('Github Apps', function() {
         })
         
         it('should return Bearer headers on success', function(complete) {
+            sandbox.stub(GithubServiceApi, "requestAccessTokens").callsFake((id, jwt, fn) => {
+                fn(token)
+            })
+
             uut.createBearer(0, function(bearer, err) {
                 expect(bearer.headers).to.be.not.undefined
                 expect(err).to.be.null
@@ -114,10 +105,11 @@ describe('Github Apps', function() {
         })
         
         it('should return error on failure', function(complete) {
-            mockResponse.ok = false
+            sandbox.stub(GithubServiceApi, "requestAccessTokens").callsFake((id, jwt, fn) => {
+                fn(null, {body: {message: "error message"}})
+            })
 
             uut.createBearer(0, function(bearer, err) {
-                expect(bearer).to.be.null
                 expect(err).to.be.not.undefined
                 complete()
             })
@@ -125,22 +117,23 @@ describe('Github Apps', function() {
     })
 
     describe('OAuth', function() {
+        var mockInstallations
+
         beforeEach(function() {
-            mockResponse = {
-                ok: true,
-                body: {
-                    installations: [
-                        {
-                            id: 1337,
-                            account: {
-                                login: "orgname",
-                                type: "type",
-                                avatar_url: "avatar"
-                            }
-                        }
-                    ]
+            mockInstallations = [
+                {
+                    id: 1337,
+                    account: {
+                        login: "orgname",
+                        type: "type",
+                        avatar_url: "avatar"
+                    }
                 }
-            }
+            ]
+
+            sandbox.stub(GithubServiceApi, "requestInstallations").callsFake((accessToken, fn) => {
+                fn(mockInstallations)
+            })
         })
 
         it('should retrieve user organizations', function(complete) {
